% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{addClassifToDb}
\alias{addClassifToDb}
\title{Insert taxonomic classification table into the taxonomic table of a Postgres database}
\usage{
addClassifToDb(
  classifTaxo,
  conn,
  tableTaxo = DBI::Id(schema = "main", table = "taxo"),
  tmpClassif = DBI::Id(schema = "main", table = "tmpclassif"),
  defRanks = DBI::Id(schema = "main", table = "def_tax_rank"),
  onConflictRank = "stop",
  onConflictParent = "stop",
  onConflictAccepted = "stop",
  onConflictGbifid = "stop",
  onConflictAuthorship = "stop"
)
}
\arguments{
\item{classifTaxo}{Taxonomic classification table, such as the one returned by extractCompleteTaxo (with arguments addLocalId=T, addAvailableAuthorship=T, getSynonyms=T)}

\item{conn}{Postgres database connection (from the RPostgres package)}

\item{tableTaxo}{name of the taxonomic table in the postgres database (for now, it is not very flexible, needs to have the fields: cd_tax, name_tax, authorship, cd_rank, cd_parent, cd_accepted and gbifid)}

\item{tmpClassif}{name of the temporary table which would receive the data from R}

\item{defRanks}{name of the taxonomic rank definition table from (with fields cd_rank, tax_rank, rank_level)}

\item{onConflictRank}{What to do in case of conflict on ranks? One of "stop", "keep" or "edit". Not implemented yet.}

\item{onConflictParent}{What to do in case of conflict on parent taxa? One of "stop", "keep" or "edit". Not implemented yet.}

\item{onConflictAccepted}{What to do in case of conflict on accepted taxa (accepted in the sense of synonymy)? One of "stop", "keep" or "edit". Not implemented yet.}

\item{onConflictGbifid}{What to do in case of conflict on gbif identificator? One of "stop", "keep" or "edit". Not implemented yet.}

\item{onConflictAuthorship}{What to do in case of conflict on authorship string? One of "stop", "keep" or "edit". Not implemented yet.}
}
\value{
returns the internal temporary table which contain all information about modifications additions and correspondences (data.frame extracted from the postgres table)
}
\description{
Insert taxonomic classification table into the taxonomic table of a Postgres database
}
