% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{addCurrentTaxoInHigherRanks}
\alias{addCurrentTaxoInHigherRanks}
\title{In the analysed Gbif object for one taxon add the taxon to the table of classification and return the classification}
\usage{
addCurrentTaxoInHigherRanks(resAnGbif)
}
\arguments{
\item{resAnGbif}{an analysed gbif object}
}
\value{
a classification table
}
\description{
In the analysed Gbif object for one taxon add the taxon to the table of classification and return the classification
}
