% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{addHigherRanks}
\alias{addHigherRanks}
\title{Add Higher ranks to a taxonomical table}
\usage{
addHigherRanks(taxo, analysedGbif, ranks = NULL, mergeOn = NA)
}
\arguments{
\item{taxo}{taxonomic table (class taxo_oneTab)}

\item{analysedGbif}{result of an analysed gbif backbone search}

\item{ranks}{ranks to add, if NULL, determined automatically}

\item{mergeOn}{which rank do we use to merge the current table with a more complete taxonomical classification table, if NA, the highest rank of taxo is used}
}
\value{
a taxonomical table with the new taxonomical ranks
}
\description{
Add Higher ranks to a taxonomical table
}
