% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{addIdentifier}
\alias{addIdentifier}
\title{Add a parameterized identifier in the taxonomic table}
\usage{
addIdentifier(
  taxo,
  field,
  nameID,
  typeID = c("external", "database", "local"),
  nameSource = NA
)
}
\arguments{
\item{taxo}{taxo_oneTab object, as created with new_taxo_oneTab}

\item{field}{field (column name of taxo) that will be considered as an identifier for a taxon}

\item{nameID}{name of the identifier (for example "gbifid")}

\item{typeID}{either "external" in case of an identifier managed by an external source, "database" in case of an identifier in a local database, or a database managed by the user, or local for a local id of a taxon}

\item{nameSource}{for instance the name of a database, or "Gbif Backbone" for the gbif backbone id of a taxon, may be NA.}
}
\value{
the taxo object with the new identifiers as an attributes
}
\description{
Add a parameterized identifier in the taxonomic table
}
