% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{analyseGbifTable}
\alias{analyseGbifTable}
\title{Analyse a table obtained from a gbif backbone search}
\usage{
analyseGbifTable(
  searched,
  tabGbif,
  rank,
  obligatory,
  expected,
  returnGbifRes = T
)
}
\description{
Analyse a table obtained from a gbif backbone search
}
