% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkGbif}
\alias{checkGbif}
\title{Check in the gbif backbone whether taxa from a particular rank in a table are erroneous and suggests corrections}
\usage{
checkGbif(
  taxo,
  rankCheck,
  higherRankCheck = higherRanks(taxo, rankCheck),
  messagesGbif = F,
  messagesOther = T,
  filters = c(kingdom = "Plantae"),
  excludeFromSuggests = c("exactMatch", "Failed"),
  epithetizeSuggests = T,
  returnGbifRes = F,
  returnAnalysedGbif = !returnGbifRes,
  returnFailed = "Failed" \%in\% excludeFromSuggests,
  returnExactMatches = F
)

checkGbif(
  taxo,
  rankCheck,
  higherRankCheck = higherRanks(taxo, rankCheck),
  messagesGbif = F,
  messagesOther = T,
  filters = c(kingdom = "Plantae"),
  excludeFromSuggests = c("exactMatch", "Failed"),
  epithetizeSuggests = T,
  returnGbifRes = F,
  returnAnalysedGbif = !returnGbifRes,
  returnFailed = "Failed" \%in\% excludeFromSuggests,
  returnExactMatches = F
)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{rankCheck}{TODO: document}

\item{higherRankCheck}{TODO: document}

\item{messagesGbif}{TODO: document}

\item{messagesOther}{TODO: document}

\item{filters}{TODO: document}

\item{excludeFromSuggests}{TODO: document}

\item{epithetizeSuggests}{TODO: document}

\item{returnGbifRes}{TODO: document}

\item{returnAnalysedGbif}{TODO: document}

\item{returnFailed}{TODO: document}

\item{returnExactMatches}{TODO: document}
}
\description{
Check in the gbif backbone whether taxa from a particular rank in a table are erroneous and suggests corrections

Check in the gbif backbone whether taxa from a particular rank in a table are erroneous and suggests corrections
}
