% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkSpace}
\alias{checkSpace}
\title{Check for irregular spaces in a taxonomic table and suggest corrections}
\usage{
checkSpace(
  taxo,
  parts = c("plot", "taxoCode", "taxonRanks", "morphoQualifiers"),
  show_ref = c(attr(taxo, "plot"), attr(taxo, "taxoCode")),
  show_space = "#"
)
}
\description{
Check for irregular spaces in a taxonomic table and suggest corrections
}
