% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkUndetermined}
\alias{checkUndetermined}
\title{Check for bad formatted undetermination qualifiers and suggest corrections}
\usage{
checkUndetermined(taxo)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}
}
\description{
Check for bad formatted undetermination qualifiers and suggest corrections
}
