% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkUndetermited}
\alias{checkUndetermited}
\title{Check for bad formatted undetermination qualifiers and suggest corrections}
\usage{
checkUndetermited(taxo)

checkUndetermited(taxo)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}
}
\description{
Check for bad formatted undetermination qualifiers and suggest corrections

Check for bad formatted undetermination qualifiers and suggest corrections
}
