% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkUnicityRankSup}
\alias{checkUnicityRankSup}
\title{Check whether taxa of a particular rank are consistently always in the same higher rank and suggest corrections}
\usage{
checkUnicityRankSup(taxo, rank = "genus", superior = "family")
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{rank}{TODO: document}

\item{superior}{TODO: document}
}
\description{
Check whether taxa of a particular rank are consistently always in the same higher rank and suggest corrections
}
