% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{correct}
\alias{correct}
\title{Correct a taxonomic table according to a table of suggested corrections}
\usage{
correct(taxo, suggested)
}
\description{
Correct a taxonomic table according to a table of suggested corrections
}
