% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{exportXlDiagnostic}
\alias{exportXlDiagnostic}
\title{Exporting the suggested corrections in an excel file}
\usage{
exportXlDiagnostic(
  suggested,
  file = file.path(getwd(), "rdsTaxValDiagnostic.xlsx"),
  overwrite = F,
  messages = T
)
}
\arguments{
\item{suggested}{data frame or list of data.frame}

\item{file}{names and path of the file to export}

\item{overwrite}{should the file be overwritten when it exists}

\item{messages}{should messages be printed}
}
\value{
returns the result of the saveInExcel
}
\description{
Exporting the suggested corrections in an excel file
}
