% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extract}
\alias{extract}
\title{Extract information from a taxonomic table}
\usage{
extract(
  taxo,
  parts = c("taxonRanks", "taxoCode", "plot", "morphoQualifiers", "comments",
    "identifiers"),
  onlyRanks = NULL,
  onlyQualifiers = NULL
)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{parts}{parts of the taxonomic information you want to extract (from the following list: "taxonRanks","taxoCode","plot","morphoQualifiers","comments","identifiers")}

\item{onlyRanks}{ranks (as defined in `taxize` package) to be extracted when "taxonRanks" in the `parts` if null all the ranks are extracted}

\item{onlyQualifiers}{qualifiers (cf_aff and/or sp_specif) to be extracted when "taxonRanks" in the `parts` if null all the ranks are extracted}
}
\description{
Extract information from a taxonomic table
}
