% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extractEpithet}
\alias{extractEpithet}
\title{Extract epithets from full name taxa}
\usage{
extractEpithet(taxon, higherTaxon = NULL)
}
\arguments{
\item{taxon}{TODO: document}

\item{higherTaxon}{TODO: document}
}
\description{
Extract epithets from full name taxa
}
