% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{formatPlots}
\alias{formatPlots}
\title{Format the taxonomical table}
\usage{
formatPlots(taxo, format = c("original", "listPlot", "oneTable"))
}
\arguments{
\item{taxo}{taxonomical table}

\item{format}{"original": take the original format as an output format (works only with the "origin" attributes of taxo_oneTable objects), "listPlot": list of taxonomical tables by plots, the name of the list is the name of the plot, "oneTable" a unique table with all the plots (plots given by a variable)}
}
\value{
see format
}
\description{
Format the taxonomical table
}
