% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getLowerTax}
\alias{getLowerTax}
\title{For each row of the taxonomic table, gives the taxon at the lower defined taxonomical rank}
\usage{
getLowerTax(taxo)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTable)}
}
\value{
a character vector of taxon names
}
\description{
For each row of the taxonomic table, gives the taxon at the lower defined taxonomical rank
}
