% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getMorphoTaxo}
\alias{getMorphoTaxo}
\title{Get morpho-taxa names from a taxonomic table}
\usage{
getMorphoTaxo(taxo)
}
\arguments{
\item{taxo}{taxo_oneTab: parameterized taxonomic table}
}
\value{
A character vector with the names of the morpho-taxa
}
\description{
Get morpho-taxa names from a taxonomic table
}
