% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getRank}
\alias{getRank}
\title{Get taxon information at a particular taxonomic rank from a taxonomic table}
\usage{
getRank(taxo, rank)
}
\description{
Get taxon information at a particular taxonomic rank from a taxonomic table
}
