% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getSuggestsFromListGbif}
\alias{getSuggestsFromListGbif}
\title{Get suggested taxa from a list of analysed tables from Gbif backbone searches}
\usage{
getSuggestsFromListGbif(
  listAnalysedGbif,
  ranks = c("family", "genus", "species", "subspecies", "variety"),
  exclude = c("exactMatch")
)
}
\arguments{
\item{listAnalysedGbif}{TODO: document}

\item{ranks}{TODO: document}

\item{exclude}{TODO: document}
}
\description{
Get suggested taxa from a list of analysed tables from Gbif backbone searches
}
