% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getSuggestsFromResGbif}
\alias{getSuggestsFromResGbif}
\title{Get suggested taxonomic information from an analysed table from a search in the Gbif backbone}
\usage{
getSuggestsFromResGbif(
  analysedGbif,
  ranks = c("family", "genus", "species", "subspecies", "variety")
)
}
\description{
Get suggested taxonomic information from an analysed table from a search in the Gbif backbone
}
