% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{higherRanks}
\alias{higherRanks}
\title{Determine which higher ranks can be found from a taxonomic table}
\usage{
higherRanks(taxo, rank, excludeEpithComponents = T)
}
\description{
Determine which higher ranks can be found from a taxonomic table
}
