% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{importXlDiagnostic}
\alias{importXlDiagnostic}
\title{Reading suggested corrections from an Excel file}
\usage{
importXlDiagnostic(file = file.path(getwd(), "rdsTaxValDiagnostic.xlsx"))
}
\arguments{
\item{file}{Excel file, modified or not, in which the suggested corrections are stored}
}
\value{
a `suggested` object such as those created by the functions `fullTaxonomicDiagnostic`, `mergeSuggest`, `checkSpace`, `checkUndetermined`, `checkUnicityRankSup`, `checkUnicityCodetax`, or `checkGbif`, but of course it really depends whether the sheets in the files are valid, no check is yet implemented in the function
}
\description{
BE CAREFUL: DO NOT READ A FILE CREATED IN ANOTHER SESSION (other time, other computer etc), the context may change the results and the row might not correspond anymore, and the package does not check yet for discrepancies...
}
