% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{mergeSuggest}
\alias{mergeSuggest}
\title{Merge various suggested correction tables into a simpler, unique one}
\usage{
mergeSuggest(
  taxo,
  lSuggest,
  descrSuggest = NULL,
  parts = c("taxoCode", "plot", "taxonRanks", "morphoQualifiers", "comments"),
  onlyRanks = NULL,
  onlyQualifiers = NULL,
  keepSuggestCol = "gbifid",
  returnEachSuggest = F,
  returnFailed = T,
  returnAnalysedGbif = T
)

mergeSuggest(
  taxo,
  lSuggest,
  descrSuggest = NULL,
  parts = c("taxoCode", "plot", "taxonRanks", "morphoQualifiers", "comments"),
  onlyRanks = NULL,
  onlyQualifiers = NULL,
  keepSuggestCol = "gbifid",
  returnEachSuggest = F,
  returnFailed = T,
  returnAnalysedGbif = T
)
}
\arguments{
\item{taxo}{taxonomic table of class taxo_oneTable (must be the initial table before applying the results)}

\item{lSuggest}{either a named list with the ordered suggested correction tables (used as recursive corrections) or the ordered names of the correction table objects in the global environment}

\item{descrSuggest}{description for each of the suggested correction table that will be used in the final suggested correction table (if null the function will use lSuggest to create descriptions)}

\item{parts}{parts of the original taxonomic table that will be shown as reference (see `extract`)}

\item{onlyRanks}{taxonomic ranks of the original taxonomic table that will be shown as reference}

\item{onlyQualifiers}{taxonomic qualifiers of the original taxonomic table that will be shown as reference}

\item{keepSuggestCol}{columns from the suggested correction table that should be kept in the final suggested correction table}

\item{returnEachSuggest}{if true all the suggested tables are included in the returned list}

\item{returnFailed}{if true the tables of failed suggested correction cases are included in the returned list}

\item{returnAnalysedGbif}{if true the analysed gbif objects are included in the returned list}
}
\value{
a list with the suggested final tables and other objects depending on the arguments given

a list with the suggested final tables and other objects depending on the arguments given
}
\description{
Merge various suggested correction tables into a simpler, unique one

Merge various suggested correction tables into a simpler, unique one
}
