% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{new_taxo_oneTab}
\alias{new_taxo_oneTab}
\title{Put the taxonomy from the RDS in one table with attributes}
\usage{
new_taxo_oneTab(
  obj,
  currentFormat = c("listPlot", "oneTable"),
  taxonRanks_names = c(family = "family", genus = "genus", species = "specificEpithet",
    infraspecies = "infraspecificEpithet"),
  taxonRanks_epithetized = c("specificEpithet", "infraspecificEpithet"),
  taxoCode = "code",
  plot = "plot",
  morphoQualifiers = c(cf_aff = "identificationQualifier", sp_specif =
    "verbatimTaxonRank"),
  comments = "comments"
)
}
\arguments{
\item{obj}{the base object with the taxonomy, may be a list with taxonomy
data.frame for each plot, or a single data.frame}

\item{currentFormat}{describe the format of obj: "listPlot" for a list of
plots and "oneTable" for a single}

\item{taxoCode}{name of the column code corresponding to code of the taxon}

\item{plot}{name of the plot (name of the permanent plot) column}

\item{family}{name of the family column}

\item{genus}{name of the genus column}

\item{epithet}{name of the specific epithet column}

\item{infraspecies_epithet}{name of the infraspecific epithet column}

\item{cf_aff}{name of the column containing the cf. or aff. expression}

\item{sp_specif}{name of the column containg information about the morphospecies}
}
\description{
Technically this creates an object with a new class inherited from data.frame
The attributes of this class will ensure that we can always understand the
structure of the object, and come back to the initial format
}
