% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{saveInExcel}
\alias{saveInExcel}
\title{Export a list of table into a multi}
\usage{
saveInExcel(file, lVar)

saveInExcel(file, lVar)
}
\arguments{
\item{file}{path and name of the excel file}

\item{lVar}{either a list of table or a vector of names of tables that will be exported as sheets in the excel files}
}
\value{
This function does not really return

This function does not really return
}
\description{
Export a list of table into a multi

Export a list of table into a multi
}
