% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{showUnicityCodetax}
\alias{showUnicityCodetax}
\title{Format and show results from the analysis of unicity of information associated with taxonomic codes}
\usage{
showUnicityCodetax(
  resUnicityCodeTax,
  type = c("code", "problems", "suggested"),
  code = NA
)
}
\description{
Format and show results from the analysis of unicity of information associated with taxonomic codes
}
