% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{suppressSuggest}
\alias{suppressSuggest}
\title{Suppress suggested corrections from a suggested correction table}
\usage{
suppressSuggest(suggested, id_suggest = NULL, row = NULL)
}
\arguments{
\item{suggested}{suggested correction table (as returned by the `check*` function)}

\item{id_suggest}{id_suggest to suppress from a suggested correction table}

\item{row}{id of rows to suppress from a suggested correction table}
}
\value{
a new suggested correction table
}
\description{
Suppress suggested corrections from a suggested correction table
}
