% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{tabTaxoFromRank}
\alias{tabTaxoFromRank}
\title{Get matrices of gbif backbone ids and names from a complete gbif classification table}
\usage{
tabTaxoFromRank(resCompleteTaxo, rank = NA)

tabTaxoFromRank(resCompleteTaxo, rank = NA)
}
\arguments{
\item{resCompleteTaxo}{obtained from `extractCompleteTaxo`}

\item{rank}{lower taxonomic rank of the table}
}
\value{
a list with 2 matrices: one of gbif backbone ids, one of canonical names

a list with 2 matrices: one of gbif backbone ids, one of canonical names
}
\description{
Get matrices of gbif backbone ids and names from a complete gbif classification table

Get matrices of gbif backbone ids and names from a complete gbif classification table
}
