% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{taxoRanks}
\alias{taxoRanks}
\title{Extract the taxonomic ranks of taxa from a taxonomic table}
\usage{
taxoRanks(taxo)

taxoRanks(taxo)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}
}
\description{
Extract the taxonomic ranks of taxa from a taxonomic table

Extract the taxonomic ranks of taxa from a taxonomic table
}
