% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{analyseGbifTable}
\alias{analyseGbifTable}
\title{Analyse a table obtained from a gbif backbone search}
\usage{
analyseGbifTable(
  searched,
  tabGbif,
  rank,
  obligatory,
  expected,
  returnGbifRes = T
)
}
\arguments{
\item{searched}{TODO: document}

\item{tabGbif}{TODO: document}

\item{rank}{TODO: document}

\item{obligatory}{TODO: document}

\item{expected}{TODO: document}

\item{returnGbifRes}{TODO: document}
}
\description{
Analyse a table obtained from a gbif backbone search
}
