% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkGbif}
\alias{checkGbif}
\title{Check in the gbif backbone whether taxa from a particular rank in a table are erroneous and suggests corrections}
\usage{
checkGbif(
  taxo,
  rankCheck,
  higherRankCheck = higherRanks(taxo, rankCheck),
  messagesGbif = F,
  messagesOther = T,
  filters = c(kingdom = "Plantae"),
  excludeFromSuggests = c("exactMatch", "Failed"),
  epithetizeSuggests = T,
  returnGbifRes = F,
  returnAnalysedGbif = !returnGbifRes,
  returnFailed = "Failed" \%in\% excludeFromSuggests,
  returnExactMatches = F
)
}
\description{
Check in the gbif backbone whether taxa from a particular rank in a table are erroneous and suggests corrections
}
