% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkSpace}
\alias{checkSpace}
\title{Check for irregular spaces in a taxonomic table and suggest corrections}
\usage{
checkSpace(
  taxo,
  parts = c("plot", "taxoCode", "taxonRanks", "morphoQualifiers"),
  show_ref = c(attr(taxo, "plot"), attr(taxo, "taxoCode")),
  show_space = "#"
)

checkSpace(
  taxo,
  parts = c("plot", "taxoCode", "taxonRanks", "morphoQualifiers"),
  show_ref = c(attr(taxo, "plot"), attr(taxo, "taxoCode")),
  show_space = "#"
)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{parts}{TODO: document}

\item{show_ref}{TODO: document}

\item{show_space}{TODO: document}
}
\description{
Check for irregular spaces in a taxonomic table and suggest corrections

Check for irregular spaces in a taxonomic table and suggest corrections
}
