% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{checkUnicityCodetax}
\alias{checkUnicityCodetax}
\title{Check whether taxonomic codes are alway associated with the same taxonomic information and suggests corrections}
\usage{
checkUnicityCodetax(taxo, noMajority = c("skip", "takeFirst", "stop"))
}
\description{
Check whether taxonomic codes are alway associated with the same taxonomic information and suggests corrections
}
