% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{correct}
\alias{correct}
\title{Correct a taxonomic table according to a table of suggested corrections}
\usage{
correct(taxo, suggested)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{suggested}{TODO: document}
}
\description{
Correct a taxonomic table according to a table of suggested corrections
}
