% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{extract}
\alias{extract}
\title{Extract information from a taxonomic table}
\usage{
extract(
  taxo,
  parts = c("taxonRanks", "taxoCode", "plot", "morphoQualifiers", "comments"),
  onlyRanks = NULL,
  onlyQualifiers = NULL
)
}
\description{
Extract information from a taxonomic table
}
