% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fullTaxonomicDiagnostic}
\alias{fullTaxonomicDiagnostic}
\title{Provide a full taxonomic diagnostic and its suggested corrections}
\usage{
fullTaxonomicDiagnostic(
  taxo,
  checks = c("spaces", "undeterminedQualifiers", "unicityInSuperiorRanks",
    "unicityCodeTax", "gbif"),
  description = c(spaces = "cleaning spaces character", undeterminedQualifiers =
    "misplaced qualifiers for undetermined taxa", unicityInSuperiorRanks =
    "unicity of \%rank\% in \%superior\%", unicityCodeTax =
    "checking for unicity of taxonomic information associated with taxonomic code", gbif
    = "Comparing \%rankCheck\% information with gbif backbone"),
  argsCheckFunction = list(),
  ...
)
}
\arguments{
\item{taxo}{TODO: document}

\item{checks}{TODO: document}

\item{description}{TODO: document}

\item{argsCheckFunction}{TODO: document}

\item{...}{TODO: document}
}
\value{
TODO: document
}
\description{
Provide a full taxonomic diagnostic and its suggested corrections
}
