% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getRank}
\alias{getRank}
\title{Get taxon information at a particular taxonomic rank from a taxonomic table}
\usage{
getRank(taxo, rank)
}
\arguments{
\item{taxo}{taxonomic table (of class taxo_oneTab)}

\item{rank}{TODO: document}
}
\description{
Get taxon information at a particular taxonomic rank from a taxonomic table
}
