% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getSuggestsFromResGbif}
\alias{getSuggestsFromResGbif}
\title{Get suggested taxonomic information from an analysed table from a search in the Gbif backbone}
\usage{
getSuggestsFromResGbif(
  analysedGbif,
  ranks = c("family", "genus", "species", "subspecies", "variety")
)

getSuggestsFromResGbif(
  analysedGbif,
  ranks = c("family", "genus", "species", "subspecies", "variety")
)
}
\arguments{
\item{analysedGbif}{TODO: document}

\item{ranks}{TODO: document}
}
\description{
Get suggested taxonomic information from an analysed table from a search in the Gbif backbone

Get suggested taxonomic information from an analysed table from a search in the Gbif backbone
}
