% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{new_taxo_oneTab}
\alias{new_taxo_oneTab}
\title{Put the taxonomy from the RDS in one table with attributes}
\usage{
new_taxo_oneTab(
  obj,
  currentFormat = c("listPlot", "oneTable"),
  taxonRanks_names = c(family = "family", genus = "genus", species = "specificEpithet",
    infraspecies = "infraspecificEpithet"),
  taxonRanks_epithetized = c("specificEpithet", "infraspecificEpithet"),
  taxoCode = "code",
  plot = "plot",
  morphoQualifiers = c(cf_aff = "identificationQualifier", sp_specif =
    "verbatimTaxonRank"),
  comments = "comments",
  identifiers = data.frame(name = character(), type = factor(levels = c("external",
    "database", "local")), nameSource = character(), field = character())
)
}
\arguments{
\item{obj}{the base object with the taxonomy, may be a list with taxonomy
data.frame for each plot, or a single data.frame}

\item{currentFormat}{describe the format of obj: "listPlot" for a list of
plots and "oneTable" for a single data.frame containing all the taxonomy}

\item{taxonRanks_names}{named character vector (names:rank as defined in taxize, content name of the column)}

\item{taxonRanks_epithetized}{names of the columns (as in taxonRanks_names) which correspond to epithets}

\item{taxoCode}{name of the column code corresponding to code of the taxon}

\item{plot}{name of the plot (name of the permanent plot) column}

\item{morphoQualifiers}{named character vector, the names are the types of qualifiers, the contents are the names of the object fields. For now it only work with cf_aff and sp_specif}

\item{comments}{name of the column containing the comments}

\item{identifiers}{data.frame with the information of identifiers for the taxa, see function AddIdentifier}
}
\description{
Technically this creates an object with a new class inherited from data.frame
The attributes of this class will ensure that we can always understand the
structure of the object, and come back to the initial format
}
