% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{saveInExcel}
\alias{saveInExcel}
\title{Export a list of table into a multi-sheet spreadsheet file}
\usage{
saveInExcel(file, lVar, overwrite = T, keepRownames = F)
}
\arguments{
\item{file}{path and name of the excel file}

\item{lVar}{either a list of table or a vector of names of tables that will be exported as sheets in the excel files}

\item{overwrite}{if the file exists, should it be overwritten?}

\item{keepRownames}{Should the rownames be exported as a column (even though the argument is TRUE, it won't have any effect when the rownames are 1:nrow)}
}
\value{
This function does not really return any important information, but it should return the result of the `openxlsx::saveWorkbook` function
}
\description{
Export a list of table into a multi-sheet spreadsheet file
}
